Start migrating ParserImpl to Kotlin...
